feat: landing page with real stats, featured specialists, request form (#819)
Replace placeholder landing with full page: hero, stats from API,
specialist carousel from /specialists/featured, services, how-it-works,
quick request form with IFNS search. Index redirects unauthenticated
users to /(public)/landing.
